abstracted program in 0 and 1
added the program as a class in a header file to include, so that it won't distract from the comments explaining the first two types of documentation (tutorials and how-to guides)
